+++ Group3/src/main/java/com/stuypulse/robot/subsystems/swerve/SimModule.java
@@ -1,40 +1,43 @@
-package com.stuypulse.robot.subsystems;
+package com.stuypulse.robot.subsystems.swerve;
 
-import com.ctre.phoenix.motorcontrol.ControlMode;
-import com.ctre.phoenix.motorcontrol.can.WPI_TalonFX;
 import com.stuypulse.robot.constants.Settings.Swerve.Turn;
+import com.stuypulse.robot.util.EncoderSim;
+import com.stuypulse.robot.util.MotorSim;
+import com.stuypulse.robot.util.MotorSim.MotorType;
 import com.stuypulse.stuylib.math.Angle;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj.DutyCycleEncoder;
 
-public class FalconModule extends SwerveModule {
+public class SimModule extends SwerveModule {
     
-    private final WPI_TalonFX drive;
-    private final WPI_TalonFX turn;
+    private final MotorSim drive;
+    private final MotorSim turn;
 
+    private final EncoderSim driveEncoder;
     private final DutyCycleEncoder turnEncoder;
     private final Angle offset;
 
-    public FalconModule(String id, Translation2d moduleOffset, int drivePort, int turnPort, int turnEncoderPort, Angle turnOffset) {
+    public SimModule(String id, Translation2d moduleOffset, int drivePort, int turnPort, int turnEncoderPort, Angle turnOffset) {
         super(id, moduleOffset);
 
-        drive = new WPI_TalonFX(drivePort);
-        turn = new WPI_TalonFX(turnPort);
+        drive = new MotorSim(MotorType.FALCON, 1, 1);
+        turn = new MotorSim(MotorType.FALCON, 1, 1);
 
+        driveEncoder = drive.getEncoder();
         turnEncoder = new DutyCycleEncoder(turnEncoderPort);
         this.offset = turnOffset;
     }
 
     @Override
     protected void setTargetVelocity(double velocity) {
-        drive.set(ControlMode.Velocity, velocity);
+        drive.set(velocity);
     }
 
     @Override
     protected void setTargetAngle(double angle) {
-        turn.set(ControlMode.Position, angle);
+        turn.set(angle);
     }
 
     @Override
@@ -48,6 +51,12 @@ public class FalconModule extends SwerveModule {
 
     @Override
     protected double getVelocity() {
-        return drive.getSelectedSensorVelocity();
+        return driveEncoder.getVelocity();
+    }
+
+    @Override
+    public void simulationPeriodic() {
+        drive.update(0.2);
+        turn.update(0.2);
     }
 }
